Add minor fixes.
Added a missing .catch statement in the Courses Component. Added a link to the  home page from the Header Component. Changed the "Cancel" link in the UpdateCourse Component to redirect to the corresponding CourseDetail page. Added fix to .catch statement in UpdateCourse Component on line 39. Added links to the home page from the NotFound, UnhandledError, and Forbidden Components